Extract generic web promotion workflow dispatch
